+++ mwcp/utils/construct/windows_structures.py
@@ -303,20 +303,33 @@ SYSTEMTIME = construct.Struct(
 
 
 # TODO: Implement _encode
-class _SystemTimeAdapter(construct.Adapter):
+class SystemTimeAdapter(construct.Adapter):
     r"""
     Adapter to convert SYSTEMTIME structured data to datetime.datetime ISO format.
 
-    >>> _SystemTimeAdapter(SYSTEMTIME).parse('\xdd\x07\t\x00\x03\x00\x12\x00\t\x00.\x00\x15\x00\xf2\x02')
+    >>> SystemTimeAdapter(SYSTEMTIME).parse(b'\xdd\x07\t\x00\x03\x00\x12\x00\t\x00.\x00\x15\x00\xf2\x02')
     '2013-09-18T09:46:21.754000'
+    >>> SystemTimeAdapter(SYSTEMTIME, tzinfo=datetime.timezone.utc).parse(b'\xdd\x07\t\x00\x03\x00\x12\x00\t\x00.\x00\x15\x00\xf2\x02')
+    '2013-09-18T09:46:21.754000+00:00
     """
+    def __init__(self, subcon, tzinfo=None):
+        """
+        :param tzinfo: Optional timezone object, default is localtime
+        :param subcon: subcon to parse SystemTime
+        """
+        super(SystemTimeAdapter, self).__init__(subcon)
+        self._tzinfo = tzinfo
+
     def _decode(self, obj, context, path):
         return datetime.datetime(
-            obj.wYear, obj.wMonth, obj.wDay, obj.wHour, obj.wMinute, obj.wSecond, obj.wMilliseconds * 1000
+            obj.wYear, obj.wMonth, obj.wDay, obj.wHour, obj.wMinute, obj.wSecond, obj.wMilliseconds * 1000,
+            tzinfo=self._tzinfo
         ).isoformat()
 
-# Hide the adapter
-SystemTime = _SystemTimeAdapter(SYSTEMTIME)
+
+# Add common helpers
+SystemTime = SystemTimeAdapter(SYSTEMTIME)
+SystemTimeUTC = SystemTimeAdapter(SYSTEMTIME, tzinfo=datetime.timezone.utc)
 
 
 EPOCH_AS_FILETIME = 116444736000000000
@@ -324,16 +337,30 @@ HUNDREDS_OF_NANOSECONDS = 10000000
 
 
 # TODO: Implement _encode
-class _FileTimeAdapter(construct.Adapter):
+class FileTimeAdapter(construct.Adapter):
     r"""
     Adapter to convert FILETIME structured data to datetime.datetime ISO format.
     Technically FILETIME is two 32-bit integers as dwLowDateTime and dwHighDateTime, but there is no need to do that
 
-    >>> _FileTimeAdapter(construct.Int64ul).parse('\x00\x93\xcc\x11\xa7\x88\xd0\x01')
+    >>> FileTimeAdapter(construct.Int64ul).parse(b'\x00\x93\xcc\x11\xa7\x88\xd0\x01')
     '2015-05-07T05:20:33'
+    >>> FileTimeAdapter(construct.Int64ul, tz=datetime.timezone.utc).parse(b'\x00\x93\xcc\x11\xa7\x88\xd0\x01')
+    '2015-05-07T09:20:33.328000+00:00'
     """
-    def _decode(self, obj, context, path):
-        return datetime.datetime.fromtimestamp((obj - EPOCH_AS_FILETIME) / HUNDREDS_OF_NANOSECONDS).isoformat()
+    def __init__(self, subcon, tz=None):
+        """
+        :param tz: Optional timezone object, default is localtime
+        :param subcon: subcon to parse FileTime
+        """
+        super(FileTimeAdapter, self).__init__(subcon)
+        self._tz = tz
 
-# Hide the adapter
-FileTime = _FileTimeAdapter(construct.Int64ul)
+    def _decode(self, obj, context, path):
+        return datetime.datetime.fromtimestamp(
+            (obj - EPOCH_AS_FILETIME) / HUNDREDS_OF_NANOSECONDS, tz=self._tz
+        ).isoformat()
+
+
+# Add common helpers
+FileTime = FileTimeAdapter(construct.Int64ul)
+FileTimeUTC = FileTimeAdapter(construct.Int64ul, tz=datetime.timezone.utc)